+++ ThirdPartyAdapters/moloco/moloco/src/test/kotlin/com/google/ads/mediation/moloco/MolocoMediationAdapterTest.kt
@@ -52,6 +52,7 @@ import com.google.android.gms.ads.mediation.rtb.SignalCallbacks
 import com.google.common.truth.Truth.assertThat
 import com.moloco.sdk.BuildConfig
 import com.moloco.sdk.publisher.Banner
+import com.moloco.sdk.publisher.BannerAdSize
 import com.moloco.sdk.publisher.CreateBannerCallback
 import com.moloco.sdk.publisher.CreateInterstitialAdCallback
 import com.moloco.sdk.publisher.CreateNativeAdCallback
@@ -60,10 +61,8 @@ import com.moloco.sdk.publisher.Initialization
 import com.moloco.sdk.publisher.InterstitialAd
 import com.moloco.sdk.publisher.MediationInfo
 import com.moloco.sdk.publisher.Moloco
-import com.moloco.sdk.publisher.Moloco.createBanner
-import com.moloco.sdk.publisher.Moloco.createBannerTablet
 import com.moloco.sdk.publisher.Moloco.createInterstitial
-import com.moloco.sdk.publisher.Moloco.createMREC
+import com.moloco.sdk.publisher.Moloco.createMolocoBanner
 import com.moloco.sdk.publisher.Moloco.createNativeAd
 import com.moloco.sdk.publisher.Moloco.createRewardedInterstitial
 import com.moloco.sdk.publisher.Moloco.getBidToken
@@ -704,9 +703,10 @@ class MolocoMediationAdapterTest {
 
       val mediationInfoCaptor = argumentCaptor<MediationInfo>()
       mockedMoloco.verify {
-        createBanner(
+        createMolocoBanner(
           mediationInfoCaptor.capture(),
           eq(TEST_AD_UNIT),
+          eq(BannerAdSize.Standard),
           eq(TEST_WATERMARK),
           createBannerCaptor.capture(),
         )
@@ -728,7 +728,13 @@ class MolocoMediationAdapterTest {
 
       adapter.loadRtbBannerAd(mediationBannerAdConfiguration, mockMediationBannerAdLoadCallback)
       mockedMoloco.verify {
-        createBanner(any(), eq(TEST_AD_UNIT), eq(TEST_WATERMARK), createBannerCaptor.capture())
+        createMolocoBanner(
+          any(),
+          eq(TEST_AD_UNIT),
+          eq(BannerAdSize.Standard),
+          eq(TEST_WATERMARK),
+          createBannerCaptor.capture(),
+        )
       }
       val capturedCallback = createBannerCaptor.firstValue
       // An example Moloco ad creation error.
@@ -756,7 +762,13 @@ class MolocoMediationAdapterTest {
       adapter.loadRtbBannerAd(mediationBannerAdConfiguration, mockMediationBannerAdLoadCallback)
 
       mockedMoloco.verify {
-        createBanner(any(), eq(TEST_AD_UNIT), eq(TEST_WATERMARK), createBannerCaptor.capture())
+        createMolocoBanner(
+          any(),
+          eq(TEST_AD_UNIT),
+          eq(BannerAdSize.Standard),
+          eq(TEST_WATERMARK),
+          createBannerCaptor.capture(),
+        )
       }
       val capturedCallback = createBannerCaptor.firstValue
       capturedCallback.invoke(/* banner= */ null, /* error= */ null)
@@ -786,9 +798,10 @@ class MolocoMediationAdapterTest {
 
       val mediationInfoCaptor = argumentCaptor<MediationInfo>()
       mockedMoloco.verify {
-        createMREC(
+        createMolocoBanner(
           mediationInfoCaptor.capture(),
           eq(TEST_AD_UNIT),
+          eq(BannerAdSize.MREC),
           eq(TEST_WATERMARK),
           createBannerCaptor.capture(),
         )
@@ -816,9 +829,10 @@ class MolocoMediationAdapterTest {
 
       val mediationInfoCaptor = argumentCaptor<MediationInfo>()
       mockedMoloco.verify {
-        createBannerTablet(
+        createMolocoBanner(
           mediationInfoCaptor.capture(),
           eq(TEST_AD_UNIT),
+          eq(BannerAdSize.Tablet),
           eq(TEST_WATERMARK),
           createBannerCaptor.capture(),
         )
@@ -844,9 +858,10 @@ class MolocoMediationAdapterTest {
 
       adapter.loadRtbBannerAd(mediationBannerAdConfiguration, mockMediationBannerAdLoadCallback)
       mockedMoloco.verify {
-        createBannerTablet(
+        createMolocoBanner(
           any(),
           eq(TEST_AD_UNIT),
+          eq(BannerAdSize.Tablet),
           eq(TEST_WATERMARK),
           createBannerCaptor.capture(),
         )
@@ -880,9 +895,10 @@ class MolocoMediationAdapterTest {
 
       adapter.loadRtbBannerAd(mediationBannerAdConfiguration, mockMediationBannerAdLoadCallback)
       mockedMoloco.verify {
-        createBannerTablet(
+        createMolocoBanner(
           any(),
           eq(TEST_AD_UNIT),
+          eq(BannerAdSize.Tablet),
           eq(TEST_WATERMARK),
           createBannerCaptor.capture(),
         )